Don't log an empty warning if pod is unready with ready containers (#34)
Luke ran into a situation where a pod's containers were all ready but
the pod's Ready condition was still False. We were logging an empty
message because we didn't have any container errors to report:

> warning: [Pod app-6734a272-cf7f4c4fd-vtr48]:

This PR changes our logic to only append the informational warning if
there's a non-nil error or if the condition actually has a message to
report.

An "unavailable" test was added to reproduce the issue. Existing tests
were extended to capture more of the messages we expect.